+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - Fix multiple reporting of ongoing arrivals (#240).
 - Set package namespace as parent of `simmer` and `wrap` environments (#241).
+- Fix arrival restart after queue drops (#257).
 
 # simmer 4.4.2